# projects/elliot-dashboard/api_spec.yaml
openapi: 3.1.0
info:
  title: Elliot Dashboard API
  description: API endpoints for Elliot's memory, knowledge, and activity dashboard
  version: 1.0.0
  contact:
    name: Agency OS
servers:
  - url: https://api.agency-os.com
    description: Production
  - url: http://localhost:8000
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /api/elliot/memory/daily:
    get:
      tags: [memory]
      summary: List daily logs
      description: Retrieve paginated list of daily memory logs
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
          description: Start date filter (inclusive)
        - name: end_date
          in: query
          schema:
            type: string
            format: date
          description: End date filter (inclusive)
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: search
          in: query
          schema:
            type: string
          description: Full-text search in accomplishments and notes
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DailyLog'
                  total:
                    type: integer
                  hasMore:
                    type: boolean
    
    post:
      tags: [memory]
      summary: Create or update daily log
      description: Upsert a daily log entry (by date)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DailyLogInput'
      responses:
        '200':
          description: Created/updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyLog'
  /api/elliot/memory/daily/{date}:
    get:
      tags: [memory]
      summary: Get specific daily log
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyLog'
        '404':
          description: Not found
  /api/elliot/memory/weekly:
    get:
      tags: [memory]
      summary: List weekly rollups
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WeeklyRollup'
...
  /api/elliot/memory/patterns:
    get:
      tags: [memory]
      summary: Get all patterns
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, resolved, archived]
            default: active
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pattern'

    post:
      tags: [memory]
      summary: Create pattern
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatternInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pattern'
...
  /api/elliot/knowledge/rules:
    get:
      tags: [knowledge]
      summary: Get all rules
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: active_only
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rule'

    put:
      tags: [knowledge]
      summary: Replace all rules
      description: Bulk update rules (syncs from RULES.md)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/RuleInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rule'
  /api/elliot/knowledge/learnings:
    get:
      tags: [knowledge]
      summary: List learnings
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: impact_level
          in: query
          schema:
            type: string
            enum: [high, medium, low]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Learning'
                  total:
                    type: integer

    post:
      tags: [knowledge]
      summary: Add learning
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LearningInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Learning'
...
  /api/elliot/knowledge/decisions:
    get:
      tags: [knowledge]
      summary: List decisions
      parameters:
        - name: outcome_status
          in: query
          schema:
            type: string
            enum: [pending, success, partial, failure]
        - name: category
          in: query
          schema:
            type: string
        - name: importance
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low]
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Decision'
                  stats:
                    type: object
                    properties:
                      total:
                        type: integer
                      pending:
                        type: integer
                      success:
                        type: integer
                      failure:
                        type: integer

    post:
      tags: [knowledge]
      summary: Log decision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Decision'
  /api/elliot/knowledge/decisions/{id}:
    get:
      tags: [knowledge]
      summary: Get decision by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Decision'

    patch:
      tags: [knowledge]
      summary: Update decision (especially outcome)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                actual_outcome:
                  type: string
                outcome_status:
                  type: string
                  enum: [pending, success, partial, failure]
                outcome_date:
                  type: string
                  format: date
                learning_extracted:
                  type: string
      responses:
        '200':
          description: Updated

  # ============================================
  # ACTIVITY ENDPOINTS
  # ============================================
  /api/elliot/activity:
    get:
      tags: [activity]
      summary: Get activity stream
      parameters:
        - name: activity_type
          in: query
          schema:
            type: string
            enum: [message, file_read, file_write, api_call, thinking, decision]
        - name: channel
          in: query
          schema:
            type: string
        - name: session_id
          in: query
          schema:
            type: string
        - name: since
          in: query
          schema:
            type: string
            format: date-time
          description: Get activities after this timestamp
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Activity'
                  latestTimestamp:
                    type: string
                    format: date-time

    post:
      tags: [activity]
      summary: Log activity (internal use)
      description: Used by Elliot/sync service to log activities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityInput'
      responses:
        '201':
          description: Created
  /api/elliot/activity/stats:
    get:
      tags: [activity]
      summary: Get activity statistics
      parameters:
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: granularity
          in: query
          schema:
            type: string
            enum: [hourly, daily, weekly]
            default: daily
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: object
                    properties:
                      totalActivities:
                        type: integer
                      totalTokens:
                        type: integer
                      activeChannels:
                        type: array
                        items:
                          type: string
                  timeSeries:
                    type: array
                    items:
                      $ref: '#/components/schemas/ActivityStats'

  # ============================================
  # SYNC ENDPOINTS
  # ============================================
  /api/elliot/sync/trigger:
    post:
      tags: [sync]
      summary: Trigger manual sync
      description: Manually trigger file → database sync
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                direction:
                  type: string
                  enum: [file_to_db, db_to_file, bidirectional]
                  default: file_to_db
                file_types:
                  type: array
                  items:
                    type: string
                    enum: [daily, weekly, patterns, rules, learnings, decisions]
                  description: Specific file types to sync (empty = all)
      responses:
        '202':
          description: Sync initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  sync_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                  message:
                    type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DailyLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        log_date:
          type: string
          format: date
        accomplishments:
          type: array
          items:
            type: string
        interactions:
          type: array
          items:
            type: object
            properties:
              channel:
                type: string
              with:
                type: string
              summary:
                type: string
              timestamp:
                type: string
        issues:
          type: array
          items:
            type: string
        notes:
          type: string
        sync_status:
          type: string
        synced_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DailyLogInput:
      type: object
      required:
        - log_date
      properties:
        log_date:
          type: string
          format: date
        accomplishments:
          type: array
          items:
            type: string
        interactions:
          type: array
          items:
            type: object
        issues:
          type: array
          items:
            type: string
        notes:
          type: string
        raw_content:
          type: string
          description: Full markdown content
...
    PatternInput:
      type: object
      required:
        - title
        - description
      properties:
        pattern_key:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
        evidence:
          type: array
          items:
            type: object

    # Knowledge Schemas
...